Light the big FFmpeg class to traits (code clean), move FFmpeg enums to enums package #11
